fix null poster path or backdrop path. set a default image.

+++ easyTMDB/src/main/java/com/mythiccosmos/easytmdb/Model/Discover/DiscoverDetailsResult.java
@@ -23,7 +23,7 @@ public class DiscoverDetailsResult {
     private Integer voteCount;
     @SerializedName("vote_average")
     @Expose
-    private Integer voteAverage;
+    private Double voteAverage;
     @SerializedName("first_air_date")
     @Expose
     private String firstAirDate;
@@ -49,7 +49,7 @@ public class DiscoverDetailsResult {
     public DiscoverDetailsResult() {
     }
 
-    public DiscoverDetailsResult(String originalName, Integer id, String name, Integer popularity, Integer voteCount, Integer voteAverage, String firstAirDate, String posterPath, List<Integer> genreIds, String originalLanguage, String backdropPath, String overview, List<String> originCountry) {
+    public DiscoverDetailsResult(String originalName, Integer id, String name, Integer popularity, Integer voteCount, Double voteAverage, String firstAirDate, String posterPath, List<Integer> genreIds, String originalLanguage, String backdropPath, String overview, List<String> originCountry) {
         this.originalName = originalName;
         this.id = id;
         this.name = name;
@@ -105,11 +105,11 @@ public class DiscoverDetailsResult {
         this.voteCount = voteCount;
     }
 
-    public Integer getVoteAverage() {
+    public Double getVoteAverage() {
         return voteAverage;
     }
 
-    public void setVoteAverage(Integer voteAverage) {
+    public void setVoteAverage(Double voteAverage) {
         this.voteAverage = voteAverage;
     }
